move to plugin from unsupported and untangle Docs from live code
