+++ docs/user/library/data/index.rst
@@ -59,13 +59,13 @@ DataStore plugins:
    property
    geopackage
    mongodb
+   csv
 
 Unsupported plugins:
 
 .. toctree::
    :maxdepth: 1
    
-   csv
    dxf
    georest
    kml
