+++ modules/plugin/csv/src/main/java/org/geotools/data/csv/parse/CSVIterator.java
