+++ modules/plugin/csv/src/main/resources/META-INF/services/org.geotools.data.DataStoreFactorySpi
