+++ modules/plugin/csv/src/test/java/org/geotools/data/csv/parse/CSVTestStrategySupport.java
