+++ modules/unsupported/pom.xml
@@ -199,7 +199,6 @@
 				</property>
 			</activation>
 			<modules>
-				<module>csv</module>
 				<module>jdbc-ng</module>
 				<module>solr</module>
 				<module>css</module>
